1) Created saix_threads.h and saix_threads.c 2) Moved all the thread workers to this file 3) Moved common defines to a newly created common_defines.h
